File pool 的最大容积，由 int 换成 long
git-svn-id: http://nutz.googlecode.com/svn/trunk@1038 423f10f2-e3a4-11dd-a6ea-a32d6b26a33d

+++ src/org/nutz/filepool/Pools.java
@@ -9,11 +9,11 @@ import java.io.File;
  */
 public class Pools {
 
-	public static File getFileById(File home, int id, String suffix) {
+	public static File getFileById(File home, long id, String suffix) {
 		return new File(getFilePathById(home, id, suffix));
 	}
 
-	public static String getFilePathById(File home, int id, String suffix) {
+	public static String getFilePathById(File home, long id, String suffix) {
 		StringBuilder sb = new StringBuilder(home.getAbsolutePath());
 		sb.append(String.format("%016X", id).replaceAll("\\p{XDigit}{2}", "/$0"));
 		if (null != suffix)
@@ -21,12 +21,12 @@ public class Pools {
 		return sb.toString();
 	}
 
-	public static int getFileId(File home, File f) {
+	public static long getFileId(File home, File f) {
 		String path = f.getAbsolutePath();
 		int pos = path.lastIndexOf('.');
 		String s = pos > 0	? path.substring(home.getAbsolutePath().length(), pos)
 							: path.substring(home.getAbsolutePath().length());
-		return Integer.parseInt(s.replaceAll("[\\\\/]", ""), 16);
+		return Long.parseLong(s.replaceAll("[\\\\/]", ""), 16);
 	}
 
 }
